pyproject.toml
[tool.poetry]
name = "imagesgrid"
version = "0.1.0"
description = ""
authors = ["maximt <vbmaxim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
